Restore test-utils
