+++ server/src/modules/users/application/use-cases/create-user/__tests__/create-user-controller.test.unit.ts
@@ -2,16 +2,12 @@ import httpMocks from 'node-mocks-http'
 import { AppError } from '../../../../../../shared/core/app-error'
 import { Result } from '../../../../../../shared/core/result'
 import { DecodedExpressRequest } from '../../../../../../shared/infra/http/routes/decoded-request'
-import { User } from '../../../../domain/entities/user'
 import { UserValueObjectErrors } from '../../../../domain/value-objects/errors'
 import { CreateUserDTO } from '../create-user-dto'
 import { CreateUserErrors } from '../create-user-errors'
-import { MockUserRepo } from '../../../../infra/repos/implementations/mock-user-repo'
-import { CreateUserController } from '../create-user-controller'
 import { CreateUserUseCase } from '../create-user-use-case'
-import { UserMap } from '../../../../mappers/user-map'
-import { UserEmail } from '../../../../domain/value-objects/user-email'
-import { UserPassword } from '../../../../domain/value-objects/user-password'
+import { createUser } from '../test-utils/create-user'
+import { setup } from '../test-utils/setup'
 
 // TODO: how to show developer these mocks are necessary when building a controller? aka must be synced with buildController()
 jest.mock('../create-user-use-case')
@@ -25,21 +21,6 @@ describe('CreateUserController', () => {
     const req = httpMocks.createRequest({ body }) as DecodedExpressRequest
     const res = httpMocks.createResponse()
     return { req, res }
-  }
-  const createUser = ({ email, password }: CreateUserDTO): User => {
-    const userEmailResult = UserEmail.create(email)
-    const userPasswordResult = UserPassword.create({ value: password, hashed: false })
-    const results = [userEmailResult, userPasswordResult] as const
-    if (!Result.resultsAllOk(results)) throw Result.getFirstError(results)
-    const userResult = User.create({ email: results[0].value, password: results[1].value })
-    if (userResult.isErr()) throw userResult.error
-    return userResult.value
-  }
-  const setup = async (users: Array<User> = []) => {
-    const userRepo = new MockUserRepo(await Promise.all(users.map(UserMap.toPersistence)))
-    const createUserUseCase = new CreateUserUseCase(userRepo)
-    const createUserController = new CreateUserController(createUserUseCase)
-    return { userRepo, createUserUseCase, createUserController }
   }
 
   test('When the CreateUserUseCase returns Ok, the CreateUserController returns 200 OK', async () => {
